ai-devops-agent/backend/requirements.txt
# === Web Framework ===
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20

# === Agent / LangGraph ===
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.7
langchain==1.2.10
langchain-core==1.2.13
langsmith==0.7.4

# === LLM Providers ===
openai==1.82.0
google-generativeai==0.8.6        # Legacy Gemini SDK — works but consider google-genai

# === Pydantic / Typing ===
pydantic==2.12.5
pydantic-core==2.41.5
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# === Git ===
gitpython==3.1.46
gitdb==4.0.12
smmap==5.0.2

# === HTTP / Networking ===
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
requests-toolbelt==1.0.0
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
h11==0.16.0

# === Testing & Linting ===
pytest==9.0.2
flake8==7.2.0                     # Added — used by test_runner + language_detector
pluggy==1.6.0
iniconfig==2.3.0

# === Environment ===
python-dotenv==1.2.1

# === Serialization ===
orjson==3.11.7
ormsgpack==1.12.2
pyyaml==6.0.3
jsonpatch==1.33
jsonpointer==3.0.0
xxhash==3.6.0
zstandard==0.25.0
uuid-utils==0.14.0

# === Google Auth (transitive from google-generativeai) ===
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1-modules==0.4.2
rsa==4.9.1

# === Crypto / Security ===
cryptography==46.0.5
cffi==2.0.0
pycparser==3.0

# === Utilities ===
anyio==4.12.1
tenacity==9.1.4
tqdm==4.67.3
packaging==26.0
pyparsing==3.3.2
pygments==2.19.2
uritemplate==4.2.0

# === Docker SDK (optional — for run_sandboxed()) ===
docker==7.1.0
